Refactor: 참조 주소 변경, 파일 명 변경, 비동기 코드 추가
