+++ server/api/src/domain/auth/service/auth.service.spec.ts
@@ -18,7 +18,7 @@ import { HistoryBuilder, LoginRequestDtoBuilder, UserBuilder, UserResponseDtoBui
 import { BadRequestException, HttpStatus, UnauthorizedException } from '@nestjs/common';
 import { UserException } from '../../../exception';
 import { LoginRequestDto } from '../dto/login-request.dto';
-import { Bcrypt } from '../../../utils/bcrypt.util';
+import { Bcrypt } from '../../../utils';
 import { User } from '../../user/user.entity';
 import { LocalDate } from 'js-joda';
 import { DevField } from '../../field/dev-field.entity';
