+++ server/socket/src/gateway/room.gateway.ts
@@ -10,7 +10,7 @@ import {
 import { Server, Socket } from 'socket.io';
 import { Logger } from '@nestjs/common';
 import { IMessage, IRoomRequest } from './room.interface';
-import { LocalDateTime, ZoneId } from '@js-joda/core';
+import { LocalDateTime} from '@js-joda/core';
 import { RoomEvent } from './room.event';
 import { RoomService } from './room.service';
